+++ src/sys/hw/mcu/board/generic/supermini_oled042.rs
@@ -6,9 +6,10 @@
 #[doc = crate::_tags!(hw namespace)]
 /// ESP32-C3 SuperMini board with a 0.42-inch OLED.
 #[doc = crate::_doc_meta!{
-    location("sys/hw/mcu/board", struct McuEsp32C3),
-    test_size_of(McuEsp32C3 = 0),
+    location("sys/hw/mcu/board", struct BoardSuperMiniOled042),
+    test_size_of(BoardSuperMiniOled042 = 0),
 }]
+/// See also [`McuEsp32C3`][crate::McuEsp32C3].
 #[derive(Debug)]
 pub struct BoardSuperMiniOled042;